requirements.txt
bs4
requests
gitpython
dnspython
lxml
